+++ contracts/src/v0.8/functions/dev/1_0_0/accessControl/TermsOfServiceAllowList.sol
@@ -33,8 +33,8 @@ contract TermsOfServiceAllowList is ITermsOfServiceAllowList, IAccessController,
   // |                     Configuration state                      |
   // ================================================================
   struct Config {
-    bool enabled;
-    address signerPublicKey;
+    bool enabled; // ═════════════╗ When enabled, access will be checked against s_allowedSenders. When disabled, all access will be allowed.
+    address signerPublicKey; // ══╝ The key pair that needs to sign the acceptance data
   }
 
   Config private s_config;